EHN simplify workflow
The current script add authorship ranking for the manuscript and update
OSF sheet with curated affiliation.
The output is passed to `neuroview_affiliations_organizer.sh` to
generate Author Arrange suited format.

The curated affiliation "affiliations_curated.tsv" should be updated
when there's new entry in osf file.
The Author ID should be continueous number starting from 1 to number of
authors, matching OSF sheet number of authors.
All information should be condfined to the header.

Ranking spreadsheet was downloaded on 20:48 pm 02.11.2021 GMT +0
The download of the file is not automated due to API aurthorization
needed.