FIX borken dependecy
I realised that I made a mistake when deleting file.
In the long term the acknowledgemnt of the book and the paper will be
different. The one prefix with prepring will be frozen after the paper
being published. Thus we need two files.

I have added extra content in the PR template to reflect that.